+++ player.py
@@ -1,3 +1,3 @@
 class Hero:
     def __init__(self):
-        self.in_battle = False
+        self.name = None
