+++ templates/Dark Quest.html
@@ -28,22 +28,20 @@ div {
   margin: 24px;
   text-align: center;
 }</style><script role="script" id="twine-user-script" type="text/twine-javascript"></script><tw-passagedata pid="1" name="Главная площадь города" tags="" position="799,345" size="100,100">ты видишь перед собой большую площадь, повсюду снуют люди. От нее есть несколько дорог: дорога на ярмарку, к алхимикам, к кузнецу, к (text-colour:blue)[лунному пику]. Куда отправишься?
-[[отправиться на ярмарку|ярмарка]]
 [[отправиться к алхимикам|алхимики]]
 [[отправиться к кузнецу|кезнец]]
 [[отправиться к лунному пику]]
-&lt;form action=&quot;http://127.0.0.1:8080/&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="2" name="ярмарка" tags="" position="462,201" size="100,100">Вы видите торговца снаряжением. Он предлогает купить вам &#39;Случайную коробку&#39;.
-[[Согласиться]]
-[[Уйти|Главная площадь города]]
-&lt;form action=&quot;http://127.0.0.1:8080/&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="3" name="алхимики" tags="" position="804,565" size="100,100">Вы видите статного алхимика в робе, предлогающего вам купить зелья. На выбор есть: (text-colour:red)[зелье здоровья]((text-colour:yellow)[1 золотая]) или (text-colour:navy)[зелье маны]((text-colour:yellow)[1 золотая]).
+&lt;form action=&quot;http://127.0.0.1:8080/main_window&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="2" name="алхимики" tags="" position="805,565" size="100,100">Вы видите статного алхимика в робе, предлогающего вам купить зелья. На выбор есть: (text-colour:red)[зелье здоровья]((text-colour:yellow)[1 золотая]) или (text-colour:navy)[зелье маны]((text-colour:yellow)[1 золотая]).
 &lt;form action=&quot;http://127.0.0.1:8080/tranzit/хилка&quot;&gt;&lt;button class=&quot;button&quot;&gt;Купить зелье здоровья&lt;/button&gt;&lt;/form&gt;
 &lt;form action=&quot;http://127.0.0.1:8080/tranzit/манка&quot;&gt;&lt;button class=&quot;button&quot;&gt;Купить зелье маны&lt;/button&gt;&lt;/form&gt;
 [[Уйти|Главная площадь города]]
-&lt;form action=&quot;http://127.0.0.1:8080/&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="4" name="кезнец" tags="" position="1175,191" size="100,100">Вы видите старого, но опытного кузнеца. Он обращает на вас внимание. Что вы купите?
+&lt;form action=&quot;http://127.0.0.1:8080/main_window&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="3" name="кезнец" tags="" position="1175,191" size="100,100">Вы видите старого, но опытного кузнеца. Он обращает на вас внимание. Что вы купите?
+[[уйти|Главная площадь города]]
 &lt;form action=&quot;armor&quot;&gt;&lt;button class=&quot;button&quot;&gt;Броню&lt;/button&gt;&lt;/form&gt;
-&lt;form action=&quot;weapon&quot;&gt;&lt;button class=&quot;button&quot;&gt;Оружие&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="5" name="отправиться к лунному пику" tags="" position="501,495" size="100,100">ничего не произошло
+&lt;form action=&quot;weapon&quot;&gt;&lt;button class=&quot;button&quot;&gt;Оружие&lt;/button&gt;&lt;/form&gt;
+&lt;form action=&quot;http://127.0.0.1:8080/main_window&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="4" name="отправиться к лунному пику" tags="" position="501,495" size="100,100">ничего не произошло
 [[Уйти|Главная площадь города]]
-&lt;form action=&quot;http://127.0.0.1:8080/&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata><tw-passagedata pid="6" name="Согласиться" tags="" position="215,292" size="100,100">Тут будет html код </tw-passagedata></tw-storydata>
+&lt;form action=&quot;http://127.0.0.1:8080/main_window&quot;&gt;&lt;button class=&quot;button&quot;&gt;На главную страницу&lt;/button&gt;&lt;/form&gt;</tw-passagedata></tw-storydata>
 <script title="Twine engine code" data-main="harlowe">"use strict";var _slicedToArray=function(){return function(e,t){if(Array.isArray(e))return e;if(Symbol.iterator in Object(e))return function(e,t){var n=[],r=!0,i=!1,a=void 0;try{for(var o,s=e[Symbol.iterator]();!(r=(o=s.next()).done)&&(n.push(o.value),!t||n.length!==t);r=!0);}catch(e){i=!0,a=e}finally{try{!r&&s.return&&s.return()}finally{if(i)throw a}}return n}(e,t);throw new TypeError("Invalid attempt to destructure non-iterable instance")}}(),_typeof="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e};function _defineProperty(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function _toConsumableArray(e){if(Array.isArray(e)){for(var t=0,n=Array(e.length);t<e.length;t++)n[t]=e[t];return n}return Array.from(e)}!function(){
 /**
  * @license almond 0.3.3 Copyright jQuery Foundation and other contributors.
