+++ templates/help.html
@@ -23,8 +23,8 @@
     а пока запомни один совет-следи за здоровьем и деньгами. Без них ты не
     проживешь тут и дня. Лутай все, что можешь, когда будешь в городе советую посетить ярмарку, там ты найдешь торговцев.
     У них ты сможешь закупиться всем, что
-    тебе надо для седующего рэйда. Если скажешь, что пришел от Гавейна, то тебе дадут скидку, возможно. И напоследок,
+    тебе надо для следующего рэйда. Если скажешь, что пришел от Гавейна, то тебе дадут скидку, возможно. И напоследок,
     посети лунный пик, когда будет возможность.</div>
-<div><form action="http://127.0.0.1:8080/"><button class="btn btn-lg btn-danger">назад</button></form></div>
+<div><form action="http://127.0.0.1:8080/main_window"><button class="btn btn-lg btn-danger">назад</button></form></div>
 </body>
 </html>
