+++ templates/test.html
@@ -8,7 +8,7 @@
 <div class="nav">
 	<a href="#">
 	<a href="http://127.0.0.1:8080/map">Карта города</a>
-	<a href="#">Персонаж</a>
+	<a href="http://127.0.0.1:8080/inventar">Персонаж</a>
 	<a href="http://127.0.0.1:8080/settings">Настройки</a>
 	<a href="http://127.0.0.1:8080/help">Помощь</a>
 </div>
